modif About.tsx : création d'un composant ProfilAndMap pour accueillir la logique de la carte et photo de profil
