+++ src/components/pages/homePage/sections/About.tsx
@@ -1,23 +1,10 @@
-import { useState } from 'react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import {
-  faMap,
-  faUser,
-  faChevronRight,
-} from '@fortawesome/free-solid-svg-icons';
+import ProfilAndMap from '../../../utility/ProfilAndMap';
 
 /** Section About
  * @description - Affiche la section About.
  * @returns {JSX.Element} - La section À propos du site.
  */
 function About(): React.JSX.Element {
-  const [visibleMap, setVisibleMap] = useState(false);
-  const [isHovered, setIsHovered] = useState(false);
-  const handleBouton = () => {
-    setVisibleMap(!visibleMap);
-    setIsHovered(false);
-  };
-
   return (
     <section id="about">
       <h2>Mon parcours de développeur</h2>
@@ -46,55 +33,7 @@ function About(): React.JSX.Element {
             continuer à évoluer et contribuer à des projets concrets.
           </p>
         </div>
-
-        <div className="w-full md:w-1/3 md:h-full flex relative">
-          <button
-            onMouseEnter={() => setIsHovered(true)}
-            onMouseLeave={() => setIsHovered(false)}
-            className={`absolute z-2 flex h-full flex-col justify-between items-start cursor-pointer bg-color-two/30 py-4 px-4 ${isHovered ? 'w-38' : 'w-12'} transition-all duration-300 ease-in-out`}
-            onClick={handleBouton}
-            aria-label={
-              visibleMap
-                ? 'Afficher la photo de profil'
-                : 'Afficher la map: tour du monde'
-            }
-          >
-            <FontAwesomeIcon
-              icon={faUser}
-              className={visibleMap ? '' : 'text-color-three'}
-            />
-
-            <div className="w-full flex justify-between items-center">
-              {isHovered && (
-                <span className=" bg-color-two text-xs px-1 py-1 rounded-md font-semibold text-center text-color-four">
-                  {visibleMap ? 'Photo de profil' : 'Afficher la map'}
-                </span>
-              )}
-              <FontAwesomeIcon
-                icon={faChevronRight}
-                className={`${isHovered ? 'rotate-180' : ''}`}
-              />
-            </div>
-
-            <FontAwesomeIcon
-              icon={faMap}
-              className={visibleMap ? 'text-color-three' : ''}
-            />
-          </button>
-
-          {visibleMap ? (
-            <div className="h-60 md:h-full w-full object-cover flex justify-center">
-              MAP
-            </div>
-          ) : (
-            <img
-              src="/images/profil.webp"
-              alt="photo Aurélien Arnaud"
-              loading="lazy"
-              className="h-60 md:h-full w-full object-cover"
-            />
-          )}
-        </div>
+        <ProfilAndMap />
       </div>
     </section>
   );
